Single resolver works with Rails 4 o 5, must require Parameters.
Keep Resolver.instance factory for easier extension in the future.
