+++ zooplankton.gemspec
@@ -23,5 +23,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "pry"
   spec.add_development_dependency "rspec", "~> 2.14"
 
-  spec.add_dependency "rails", ">= 5.0"
+  spec.add_dependency "rails", ">= 4.0"
 end
